+++ frontend/javascripts/prototypes/new_volume_architecture/bucket_voxel_mask.ts
