+++ frontend/javascripts/test/prototypes/new_volume_architecture/bucket_voxel_mask.spec.ts
